requirements.txt
fastapi
uvicorn
pandas
scikit-learn
joblib
python-dotenv
streamlit
requests
google-generativeai